Day 7 Part 2: What an adventure
Original implementation I tried recursion, but then it ended up being
too slow.

So then I tried using a queue, and it was still too slow.

I learned to use cProfile as well, which was pretty cool.
`python -m cProfile myscript.py`
It measures the times of everything as well as the different function
calls.

The issue is that for part 2 we would be calculating too many items
using recursion / queue. I was trying to calculate everything. In part
1, we can skip calculating every single beam since we only care about
how often things split. But for part 2 I was trying to calculate every
beam during a split, and that's just too many dang beams, because the
final answer was in the 10 trillion range. So I would have to calculate
the beam for 10 trillion items.

So after reading these two reddit threads, I followed their approach for
a simpler solution:
https://www.reddit.com/r/adventofcode/comments/1pgb377/2025_day_7_part_2_hint/
https://www.reddit.com/r/adventofcode/comments/1pgnmou/2025_day_7_lets_visualize/

Essentially we just keep track of the *number* of beams, but we don't
have to actually process each beam. Much more elegant and much easier to
compute!