Day 9 part 2
First pass: Iterate through the entire rectangle and see if each part of
the rectangle was within bounds. Of course this fails though because the
rectangles are HUUUUUGE.

Iterating through the entire rectangle would take forever.

Okay so then I came up with this instead, and it took awhile to
implement because of bugs and stuff.

General algo:
1. Get all the areas and sort by biggest
2. Go through each one to determine if it's within bounds
3. I basically combine the perimeters to see if there's a combined top, bottom,
left, right perimeters that would be bigger than the rectangle I'm trying to
create.